Refactor of redux components, styles and corrections that Santi required
