--- src/App/styles.js
@@ -1,12 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-const styles = StyleSheet.create({
-  footerStyle: {
-    position: 'absolute',
-    left: 0,
-    right: 0,
-    bottom: 0
-  }
-});
-
-export default styles;
