--- src/Components/CardSection/styles.js
@@ -1,20 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-const styles = StyleSheet.create({
-  containerStyle: {
-    borderBottomWidth: 1,
-    padding: 5,
-    backgroundColor: '#FFF',
-    justifyContent: 'flex-start',
-    flexDirection: 'row',
-    alignItems: 'center',
-    borderColor: '#ddd',
-    position: 'relative'
-  },
-  checkedStyle: {
-    backgroundColor: '#dddddd',
-    borderColor: '#b1b1b1'
-  }
-});
-
-export default styles;
