--- src/Components/CheckBox/styles.js
@@ -1,23 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-const styles = StyleSheet.create({
-  uncheckedCheckBoxStyle: {
-    width: 30,
-    height: 30,
-    borderColor: '#000',
-    borderRadius: 2,
-    borderWidth: 2,
-    marginRight: 10,
-    padding: 3,
-    backgroundColor: '#FFF'
-  },
-  checkedCheckBoxStyle: {
-    borderColor: '#000',
-    flex: 1,
-    borderRadius: 1,
-    borderWidth: 1,
-    backgroundColor: '#000'
-  }
-});
-
-export default styles;
