--- src/Components/DeleteSelectedButton/styles.js
@@ -1,10 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-const styles = StyleSheet.create({
-  textStyle: {
-    fontSize: 18,
-    color: 'red'
-  }
-});
-
-export default styles;
