--- src/Components/Footer/styles.js
@@ -1,17 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-const styles = StyleSheet.create({
-  footerStyle: {
-    alignItems: 'center',
-    padding: 20,
-    backgroundColor: '#FFF',
-    borderColor: '#ddd',
-    borderWidth: 0.5
-  },
-  textStyle: {
-    fontSize: 18,
-    color: 'red'
-  }
-});
-
-export default styles;
