--- src/Components/IconButton/styles.js
@@ -1,11 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-const styles = StyleSheet.create({
-  iconButtonStyle: {
-    justifyContent: 'center',
-    marginRight: 8,
-    marginLeft: 5
-  }
-});
-
-export default styles;
