--- src/Components/Input/styles.js
@@ -1,11 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-const styles = StyleSheet.create({
-  textInputStyle: {
-    flex: 1,
-    fontSize: 18,
-    paddingLeft: 15
-  }
-});
-
-export default styles;
