--- src/Components/NavigationButton/styles.js
@@ -1,17 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-const styles = StyleSheet.create({
-  buttonStyle: {
-    flex: 1,
-    alignItems: 'center',
-    padding: 20,
-    backgroundColor: '#FFF',
-    borderColor: '#ddd',
-    borderWidth: 0.5
-  },
-  textStyle: {
-    fontSize: 18
-  }
-});
-
-export default styles;
