--- src/Components/Title/index.js
@@ -1,15 +1,0 @@
-import React from 'react';
-import { Text, View, Button } from 'react-native';
-
-import styles from './styles';
-
-const Title = () => {
-  const { textStyle, viewStyle } = styles;
-  return (
-    <View style={viewStyle}>
-      <Text style={textStyle}>Todo List</Text>
-    </View>
-  );
-};
-
-export default Title;
