--- src/Components/Title/styles.js
@@ -1,17 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-const styles = StyleSheet.create({
-  viewStyle: {
-    backgroundColor: '#87ceeb',
-    justifyContent: 'center',
-    alignItems: 'center',
-    padding: 15,
-    flex: 1
-  },
-  textStyle: {
-    fontSize: 20,
-    color: '#FFF'
-  }
-});
-
-export default styles;
