--- src/Components/Todo/styles.js
@@ -1,15 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-const styles = StyleSheet.create({
-  footerStyle: {
-    position: 'absolute',
-    left: 0,
-    right: 0,
-    bottom: 0
-  },
-  viewStyle: {
-    flex: 1
-  }
-});
-
-export default styles;
