+++ src/components/CardSection/styles.js
@@ -1,0 +1,18 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    borderBottomWidth: 1,
+    padding: 5,
+    backgroundColor: '#FFF',
+    flexDirection: 'row',
+    alignItems: 'center',
+    borderColor: '#ddd'
+  },
+  checked: {
+    backgroundColor: '#dddddd',
+    borderColor: '#b1b1b1'
+  }
+});
+
+export default styles;
