+++ src/components/DeleteSelectedButton/styles.js
@@ -1,0 +1,10 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  text: {
+    fontSize: 18,
+    color: 'red'
+  }
+});
+
+export default styles;
