+++ src/components/DummyScreen/styles.js
