+++ src/components/Footer/styles.js
@@ -1,0 +1,17 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  footer: {
+    alignItems: 'center',
+    padding: 20,
+    backgroundColor: '#FFF',
+    borderColor: '#ddd',
+    borderWidth: 0.5
+  },
+  text: {
+    fontSize: 18,
+    color: 'red'
+  }
+});
+
+export default styles;
