+++ src/components/IconButton/styles.js
@@ -1,0 +1,11 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  iconButton: {
+    justifyContent: 'center',
+    marginRight: 8,
+    marginLeft: 5
+  }
+});
+
+export default styles;
