+++ src/components/Input/styles.js
@@ -1,0 +1,11 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  textInput: {
+    flex: 1,
+    fontSize: 18,
+    paddingLeft: 15
+  }
+});
+
+export default styles;
