+++ src/components/NavigationButton/styles.js
@@ -1,0 +1,17 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  button: {
+    flex: 1,
+    alignItems: 'center',
+    padding: 20,
+    backgroundColor: '#FFF',
+    borderColor: '#ddd',
+    borderWidth: 0.5
+  },
+  textStyle: {
+    fontSize: 18
+  }
+});
+
+export default styles;
