+++ src/components/Title/styles.js
@@ -1,0 +1,17 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  view: {
+    backgroundColor: '#87ceeb',
+    justifyContent: 'center',
+    alignItems: 'center',
+    padding: 15,
+    flex: 1
+  },
+  text: {
+    fontSize: 20,
+    color: '#FFF'
+  }
+});
+
+export default styles;
