+++ src/components/Todo/styles.js
@@ -1,0 +1,15 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  footer: {
+    position: 'absolute',
+    left: 0,
+    right: 0,
+    bottom: 0
+  },
+  viewStyle: {
+    flex: 1
+  }
+});
+
+export default styles;
